Move assertConfigsAreEqual to utility class
