+++ spring-security/src/test/java/com/sap/cloud/security/spring/config/XsuaaServiceConfigurationTest.java
@@ -6,9 +6,7 @@
 package com.sap.cloud.security.spring.config;
 
 import com.sap.cloud.environment.servicebinding.SapVcapServicesServiceBindingAccessor;
-import com.sap.cloud.security.config.OAuth2ServiceConfiguration;
 import com.sap.cloud.security.config.ServiceBindingEnvironment;
-import com.sap.cloud.security.config.ServiceConstants;
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -21,7 +19,6 @@ import org.springframework.context.annotation.PropertySource;
 import java.io.IOException;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
 @SpringBootTest(classes = { SingleXsuaaConfigurationFromFile.class })
@@ -40,16 +37,9 @@ class XsuaaServiceConfigurationTest {
 
 	@Test
 	void configuresXsuaaServiceConfiguration() {
-		assertConfigsAreEqual(xsuaaConfig, env.getXsuaaConfiguration());
+		ConfigurationUtil.assertConfigsAreEqual(xsuaaConfig, env.getXsuaaConfiguration());
     }
 
-	static void assertConfigsAreEqual(XsuaaServiceConfiguration xsuaaConfig, OAuth2ServiceConfiguration oauthConfig) {
-		assertEquals(oauthConfig.getClientId(), xsuaaConfig.getClientId());
-		assertEquals(oauthConfig.getClientSecret(), xsuaaConfig.getClientSecret());
-		assertEquals(oauthConfig.getProperty(ServiceConstants.XSUAA.UAA_DOMAIN), xsuaaConfig.getProperty(ServiceConstants.XSUAA.UAA_DOMAIN));
-		assertEquals(oauthConfig.getProperty(ServiceConstants.XSUAA.APP_ID), xsuaaConfig.getProperty(ServiceConstants.XSUAA.APP_ID));
-		assertEquals(oauthConfig.getProperty(ServiceConstants.SERVICE_PLAN), xsuaaConfig.getPlan());
-	}
 }
 
 @Configuration
